+++ libs/openFrameworks/video/ofVideoPlayer.h
@@ -77,7 +77,7 @@ class ofVideoPlayer : public ofBaseVideo{
 		//the anchor is the point the image is drawn around.
 		//this can be useful if you want to rotate an image around a particular point.
         void				setAnchorPercent(float xPct, float yPct);	//set the anchor as a percentage of the image width/height ( 0.0-1.0 range )
-        void				setAnchorPoint(int x, int y);				//set the anchor point in pixels
+        void				setAnchorPoint(float x, float y);				//set the anchor point in pixels
         void				resetAnchor();								//resets the anchor to (0, 0)
 
 		void 				setPaused(bool bPause);
